+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 v0.4.0 (in development)
 -----------------------
+- Split off the command-line programs from
+  [`javaproperties`](https://github.com/jwodder/javaproperties) into a separate
+  package,
+  [`javaproperties-cli`](https://github.com/jwodder/javaproperties-cli)
 
 v0.3.0 (2017-04-13)
 -------------------